NeighborlyAPI/requirements.txt
azure-functions==1.9.0
azure-cosmos==4.2.0
azure-cli-core==2.34.1
pymongo==4.0.2
Flask==1.1.2
requests==2.23.0